Install the app icon to the hicolor theme
From the standard:
In order to have a place for third party applications to install their
icons there should always exist a theme called "hicolor". The data for
the hicolor theme is available for download at:
http://www.freedesktop.org/software/icon-theme/.
Implementations are required to look in the "hicolor" theme if an icon
was not found in the current theme.