Cleanup and fixes
Fixes https://github.com/lxqt/screengrab/issues/48 and https://github.com/lxqt/screengrab/issues/49

(1) The screenshot settings of the main window are saved/restored (as in Spectacle). They are shown only when relevant to the chosen screenshot type.

(2) The item "Previous selecttion" is changed to "Last selected area" to be unambiguous. It means a screenshot inside the previously selected rectangle. Now, the selected rectangle is also saved on quitting and restored with the next session.

(3) Image format on saving is fixed.

(4) The special text of the spinbox ("None") is set in the ui file, not just when it has a zero value (that was a bad code).

(5) Also, the reading and writing of settings is simplified so that global config files could be used.