Handle invalid window screenshots
Fixes https://github.com/lxqt/screengrab/issues/51.

The following checks are made to make a valid window screenshot:

(1) The window is active and isn't screengrab's;
(2) It's on the current desktop; and
(3) It isn't a desktop or panel/dock.

If any of these conditions isn't fulfilled (which is possible according to my tests), a fullscreen shot is taken instead of a window screenshot or instead of exiting.