Several fixes, especially for the tray icon
Fixes https://github.com/lxqt/screengrab/issues/63 and https://github.com/lxqt/screengrab/issues/61

(1) Several issues are fixed in the tray icon. Adding their long list to this comment wouldn't be helpful.

(2) The height of the About tab-bar shouldn't be enforced; otherwise, it'll interfer with the active widget style. Either don't do hard-coded styling or do it completely! There was no need to the latter here.

(3) The theme icon (if any) should always be preferred to the hard-coded icon (except for special cases that aren't related to this app).

(4) A wrong path is corrected in `aboutwidget.ui`.