Fixed screen area and window modes with scale factors > 1
Long story short, although the screen area widget covers the whole screen, the widget coordinates are different from the pixmap coordinates when scaling is enabled.

Also:

 1. In the window mode, "KWindowInfo::frameGeometry" doesn't consider the screen's device pixel ratio. So, it's taken into account by the patch.
 2. The device pixel ratio wasn't considered in grabbing the mouse cursor image either. That's corrected.
 3. Some old mistakes are corrected in the zooming code.

As far as I tested it, the patch worked fine with the scale factors 2 and, especially, 1.5 — the latter guarantees that it should work with all scale factors.

Closes https://github.com/lxqt/lxqt/issues/1814