Cleaned up GUI/code and added an option for Wayland (#401)
* Cleaned up GUI/code and added an option for Wayland

The fitting option is made hidden and the window screenshot is removed under Wayland.

Also, a Wayland-only option is added to the Advanced page for remembering the last selected screen in the next session. It's unchecked by default because it may be confusing for some users.

NOTE: The screen combo-box is needed only because there's no way of knowing the screen with cursor under Wayland. At the current pace, this should be fixed by Wayland devs in the next century. Then I'll remove that combo-box and its corresponding option.

* Make sure the displayed image fits the first window

* Removed 2 redundant slots

All changes in the Options dialog should be applied explicitly, not on the fly.

* Fixed a miscalculation in tip's position

* Removed toolbar context menu

* Fixed a crash on quitting from tray

The crash happened when the window was minimized to tray